// services/UI/src/app/api/model/restaurant.ts
/**
 * Mealist API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Restaurant { 
    id?: string;
    name: string;
    address: string;
    city: string;
    description?: string;
    wifiName?: string;
    wifiPassword?: string;
    instagramUrl?: string;
    vkUrl?: string;
    facebookUrl?: string;
    tripAdvisorUrl?: string;
}